One commit per save
